Remove log record from fixture names?
